consolidate deprecated classes; get rid of RobotEntity
