+++ src/main/java/com/marginallyclever/robotoverlord/robots/skycam/SkycamCommand.java
@@ -2,7 +2,7 @@ package com.marginallyclever.robotoverlord.robots.skycam;
 
 import com.marginallyclever.convenience.StringHelper;
 import com.marginallyclever.robotoverlord.Entity;
-import com.marginallyclever.robotoverlord.entities.PoseEntity;
+import com.marginallyclever.robotoverlord.robots.PoseEntity;
 import com.marginallyclever.robotoverlord.parameters.DoubleParameter;
 import com.marginallyclever.robotoverlord.swinginterface.componentmanagerpanel.ComponentPanelFactory;
