+++ src/main/java/com/marginallyclever/robotoverlord/robots/stewartplatform/linear/LinearStewartPlatform1.java
@@ -2,7 +2,7 @@ package com.marginallyclever.robotoverlord.robots.stewartplatform.linear;
 
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.MatrixHelper;
-import com.marginallyclever.robotoverlord.entities.ShapeEntity;
+import com.marginallyclever.robotoverlord.robots.ShapeEntity;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Vector3d;
