+++ src/main/java/com/marginallyclever/robotoverlord/robots/stewartplatform/linear/LinearStewartPlatformCore.java
@@ -4,7 +4,7 @@ package com.marginallyclever.robotoverlord.robots.stewartplatform.linear;
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.*;
 import com.marginallyclever.robotoverlord.components.MaterialComponent;
-import com.marginallyclever.robotoverlord.entities.PoseEntity;
+import com.marginallyclever.robotoverlord.robots.PoseEntity;
 import com.marginallyclever.robotoverlord.parameters.BooleanParameter;
 import com.marginallyclever.robotoverlord.parameters.DoubleParameter;
 import com.marginallyclever.robotoverlord.parameters.RemoteParameter;
